Add pseudocode

+++ internal/oidc/token-handler.go
@@ -18,7 +18,7 @@ import (
 type TokenHandlerResponse struct {
 	AccessToken  string `json:"access_token"`
 	TokenType    string `json:"token_type"`
-	ExpiresIn    uint32 `json:"expires_in"`
+	ExpiresIn    int64  `json:"expires_in"`
 	RefreshToken string `json:"refresh_token"`
 	IDToken      string `json:"id_token"`
 }
